+++ RELEASENOTES.md
@@ -77,6 +77,9 @@
         ([#3088](https://github.com/androidx/media/issues/3088)).
     *   MP3: Ignore Xing data length if it's longer than the known stream length
         ([#3117](https://github.com/androidx/media/issues/3117)).
+    *   MP3: Use gapless-aware durations from Xing/Info headers while keeping
+        raw durations for bitrate calculations
+        ([#3183](https://github.com/androidx/media/issues/3183)).
     *   Ignore `av1C` data with unsupported version.
     *   MP4: Add support for big-endian floating point PCM in `fpcm` boxes.
     *   Matroska: Parse chapter info to `Chapter` entries in a track's